publisher/requirements.txt
docker
influxdb_client
python-dotenv
requests